% Auto-generated: do not edit by hand
\name{spillthebeansThreejs}

\alias{spillthebeansThreejs}

\title{SpillthebeansThreejs component}

\description{

}

\usage{
spillthebeansThreejs(id=NULL, beans=NULL, selected=NULL)
}

\arguments{
\item{id}{The ID used to identify this component in Dash callbacks.}

\item{beans}{The beans!}

\item{selected}{The selected bean.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

