% Auto-generated: do not edit by hand
\name{spillthebeansThreejs}

\alias{spillthebeansThreejs}

\title{SpillthebeansThreejs component}

\description{

}

\usage{
spillthebeansThreejs(id=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify the container for the IFC viewer component.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

